chore(refactor): split pivot/types and pivot/helpers (refs #47)
