+++ test/unit/pivot/mapping.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { isGeometryType, isValidPropertyType } from '@/pivot/types';
+import { isGeometryType, isValidPropertyType } from '@/pivot/helpers';
 
 describe('isValidPropertyType', () => {
   it('should return true for "string"', () => {
